+++ BitmovinAnalyticsCollector/Classes/Collector/stateMachines/StateMachine.swift
@@ -12,8 +12,6 @@ public class StateMachine {
     
     //tracked player times
     private(set) var stateEnterTimestamp: Int64 = 0
-    public var potentialSeekStart: Int64 = 0
-    public var potentialSeekVideoTimeStart: CMTime?
     var startupTime: Int64 = 0
     private(set) var videoTimeStart: CMTime?
     internal var videoTimeEnd: CMTime?
@@ -76,19 +74,23 @@ public class StateMachine {
     }
 
     public func transitionState(destinationState: PlayerState, time: CMTime?) {
+        transitionState(destinationState: destinationState, playerTime: time, enterTimestamp: Date().timeIntervalSince1970Millis)
+    }
+    
+    private func transitionState(destinationState: PlayerState, playerTime: CMTime?, enterTimestamp overrideEnterTimestamp: Int64?) {
         let performTransition = checkUnallowedTransitions(destinationState: destinationState)
 
         if performTransition {
             print("[StateMachine] Transitioning from state \(state) to \(destinationState)")
-            let currentTimestamp = Date().timeIntervalSince1970Millis
-            videoTimeEnd = time
+            let usedEnterTimestamp = overrideEnterTimestamp ?? Date().timeIntervalSince1970Millis
+            videoTimeEnd = playerTime
             
             // Get the time spend in the current state
-            let duration = currentTimestamp - stateEnterTimestamp
+            let duration = usedEnterTimestamp - stateEnterTimestamp
             state.onExit(stateMachine: self, duration: duration, destinationState: destinationState)
             
             state = destinationState
-            stateEnterTimestamp = currentTimestamp
+            stateEnterTimestamp = usedEnterTimestamp
             videoTimeStart = videoTimeEnd
             state.onEntry(stateMachine: self)
         }
@@ -110,6 +112,10 @@ public class StateMachine {
     
     public func playing(time: CMTime?) {
         transitionState(destinationState: .playing, time: time)
+    }
+    
+    public func seek(time: CMTime?, overrideEnterTimestamp: Int64? = nil) {
+        transitionState(destinationState: .seeking, playerTime: time, enterTimestamp: overrideEnterTimestamp)
     }
     
     public func videoQualityChange(time: CMTime?) {
@@ -219,13 +225,6 @@ public class StateMachine {
         }
         
         return true
-    }
-
-    public func confirmSeek() {
-        stateEnterTimestamp = potentialSeekStart
-        videoTimeStart = potentialSeekVideoTimeStart
-        potentialSeekStart = 0
-        potentialSeekVideoTimeStart = CMTime.zero
     }
 
     func enableHeartbeat() {